requirements.txt
Flask
flask-cors
couchdb
kafka-python